Split fact promotion owner helpers
